+++ relay-event-schema/src/protocol/span.rs
@@ -91,6 +91,12 @@ pub struct Span {
     #[metastructure(skip_serialization = "empty")]
     pub _metrics_summary: Annotated<MetricsSummary>,
 
+    /// Platform identifier.
+    ///
+    /// See [`Event::platform`].
+    #[metastructure(skip_serialization = "empty")]
+    pub platform: Annotated<String>,
+
     // TODO remove retain when the api stabilizes
     /// Additional arbitrary fields for forwards compatibility.
     #[metastructure(additional_properties, retain = "true", pii = "maybe")]
@@ -107,6 +113,7 @@ impl From<&Event> for Span {
             start_timestamp: event.start_timestamp.clone(),
             timestamp: event.timestamp.clone(),
             measurements: event.measurements.clone(),
+            platform: event.platform.clone(),
             ..Default::default()
         };
 
@@ -257,6 +264,11 @@ pub struct SpanData {
     #[metastructure(field = "thread.name")]
     pub thread_name: Annotated<Value>,
 
+    /// Origin Transaction name of the span.
+    ///
+    /// For INP spans, this is the route name where the interaction occurred.
+    pub transaction: Annotated<String>,
+
     /// Name of the UI component (e.g. React).
     #[metastructure(field = "ui.component_name")]
     pub ui_component_name: Annotated<Value>,
@@ -300,7 +312,7 @@ impl Getter for SpanData {
             "thread\\.name" => self.thread_name.value()?.into(),
             "ui\\.component_name" => self.ui_component_name.value()?.into(),
             "url\\.scheme" => self.url_scheme.value()?.into(),
-
+            "transaction" => self.transaction.as_str()?.into(),
             _ => {
                 let escaped = path.replace("\\.", "\0");
                 let mut path = escaped.split('.').map(|s| s.replace('\0', "."));
@@ -482,6 +494,7 @@ mod tests {
                     ],
                 },
             ),
+            platform: ~,
             other: {},
         }
         "###);
@@ -546,6 +559,7 @@ mod tests {
             server_address: ~,
             http_response_status_code: ~,
             thread_name: ~,
+            transaction: ~,
             ui_component_name: ~,
             url_scheme: ~,
             other: {