+++ py/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - This release requires Python 3.10 or later. There are no intentionally breaking changes included in this release, but we stopped testing against Python 3.9.
+- Fix: accept transactions with unfinished spans. ([#3162](https://github.com/getsentry/relay/pull/3162))
 
 ## 0.8.45